Make the lisplexer function into a class.
Convert the lisp lexer function into a class with an explicit build method. This
makes it easier to refactor things, as we can reference the functions inside the
class, where before it was impossible to reference the functions inside the
lisplexer function.

Also explicitly pass the lexer to the parse method of the parser in repl.py.
This reduces the amount of magic involved, and makes it easier to see where yacc
is getting it's references from.