Made the lispparser function into a class.
The build method of the class returns the parse object, not the parse method,
which was what the lispparser function returned before. This means that to parse
a string, you must call PyClojureParser().build().parse("foobar"). However, it
means we have the entire object, instead of just the parse function, which
allows a greater degree of flexebility.