Added support for a wider range of numeric literals.
At lexing, integers and floats are found using the same regex. This is due to
the fact that finding two regexes for integers and floats that will coexist is
very hard; ply's lexing seems to fail here, as using a valid regex for integers
and a valid regex for floats will not work, so I have combined them into a
NUMBER token, which will then be changed into either a FLOAT or INTEGER token.

However, this creates the problem that the lexer tokens no longer equal the
parser tokens. For the moment I have just manually removed NUMBER from the
parser tokens and added INTEGER and FLOAT, but I can see no need to
differentiate between the two when parsing, so they could be merged.